requirements.txt
pynest-api
fastapi
uvicorn[standard]
jinja2
python-multipart
qrcode[pil]
pillow
python-magic
aiofiles